ml_service/requirements.txt
# PyTorch and Vision
torch>=2.5.0
torchvision>=0.20.0

# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
Pillow==10.1.0

# Scientific Computing
numpy==1.26.2

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0

# Training Utilities
tqdm==4.66.1
matplotlib==3.8.2
scikit-learn==1.3.2
